requirements.txt
# Core
selenium>=4.15.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
waitress>=2.1.2

# Utilidades
python-dotenv>=1.0.0